updated create && remove post key functions to contain user authentication process; also added create portfolio, which allows a user to create a new portfolio page

+++ flask_backend/tapa_env/app.py
@@ -107,31 +107,65 @@ def add_post_key(username,password,post_key):
     for repl in get_post_key_repl:
         post_keys.append(repl[0])
     
-    if post_key not in post_keys:
-        add_post_key_query = f"INSERT INTO user_post_key(user_id,post_key) VALUES({user_id},\"{post_key}\")"
-        print("add post key query : ", add_post_key_query)
-        tapa_cursor.execute(add_post_key_query)
-        db_conn.commit()
-        return f"Successfully added post key associated with user {user_id}"
-    else:
-        return "This post key already exists. Try create a different post key."
+    if user_id >= 0:
+        if post_key not in post_keys:
+            add_post_key_query = f"INSERT INTO user_post_key(user_id,post_key) VALUES({user_id},\"{post_key}\")"
+            print("add post key query : ", add_post_key_query)
+            tapa_cursor.execute(add_post_key_query)
+            db_conn.commit()
+            return f"Successfully added post key associated with user {user_id}"
+        else:
+            return "This post key already exists. Try create a different post key."
+    elif user_id == -1:
+        return "User authentication failed"
+    elif user_id == -2:
+        return "No account with such username exists"
     
 @app.route('/deletePostKey/<string:username>/<string:password>/<string:post_key>')
 def delete_post_key(username,password,post_key):
     user_id = authenticate_user_account(username,password)
-    get_users_postkeys_query = f"SELECT post_key FROM user_post_key WHERE user_id=\"{user_id}\""
+    get_users_postkeys_query = f"SELECT post_key FROM user_post_key WHERE user_id=\"{username}\""
     tapa_cursor.execute(get_users_postkeys_query)
     get_post_key_repl = tapa_cursor.fetchall()
     post_keys = []
     for repl in get_post_key_repl:
         post_keys.append(repl[0])
     
-    if post_key in post_keys:
-        delete_post_key_query = f"DELETE FROM user_post_key WHERE post_key=\"{post_key}\""
-        tapa_cursor.execute(delete_post_key_query)
-        db_conn.commit()
-        return "Successfully removed post key"
-    else:
-        return "This post key does not exist."
+    if user_id >= 0:
+        if post_key in post_keys:
+            delete_post_key_query = f"DELETE FROM user_post_key WHERE post_key=\"{post_key}\""
+            tapa_cursor.execute(delete_post_key_query)
+            db_conn.commit()
+            return "Successfully removed post key"
+        else:
+            return "This post key does not exist."
+    elif user_id == -1:
+        return "User authentication failed"
+    elif user_id == -2:
+        return "No account with such username exists"
+    
+@app.route('/createPortfolio/<string:username>/<string:password>/<string:portfolio_name>/<string:introduction>/<string:phone>/<string:email>')
+def create_portfolio(username,password,portfolio_name,introduction,phone,email):
+    user_id = authenticate_user_account(username,password)
+    get_existing_portfolio_query = f"SELECT portfolio_link FROM user_to_portfolio WHERE user_id>0"
+    tapa_cursor.execute(get_existing_portfolio_query)
+    existing_portfolios = [row[0] for row in tapa_cursor.fetchall()]
+    if user_id >= 0:
+        if portfolio_name not in existing_portfolios:
+            create_new_portfolio_query = f"INSERT INTO portfolio(portfolio_link,introduction,contact_phone,contact_email) VALUES(\"{portfolio_name}\",\"{introduction}\",\"{phone}\",\"{email}\")"
+            tapa_cursor.execute(create_new_portfolio_query)
+            db_conn.commit()
+            
+            link_user_to_portfolio_query = f"INSERT INTO user_to_portfolio(user_id,portfolio_link) VALUES(\"{user_id}\",\"{portfolio_name}\")"
+            tapa_cursor.execute(link_user_to_portfolio_query)
+            db_conn.commit()
+            return f"Successfully created a new portfolio for user \"{username}\""
+        else:
+            return "This portfolio name has already been used"
+    elif user_id == -1:
+        return "User authentication failed"
+    elif user_id == -2:
+        return "No account with such username exists"
+    
 if __name__=="__main__":
     app.run(debug=True)
